requirements.txt
wheel
setuptools
PyYAML
Pillow
